+++ small_text/integrations/transformers/classifiers/classification.py
@@ -19,6 +19,7 @@ from small_text.integrations.transformers.classifiers.base import (
     get_default_model_loading_strategy
 )
 
+from small_text.training.model_selection import NoopModelSelection
 from small_text.utils.logging import verbosity_logger, VERBOSITY_MORE_VERBOSE
 from small_text.utils.system import get_tmp_dir_base
 
@@ -326,7 +327,7 @@ class TransformerBasedClassification(TransformerBasedEmbeddingMixin, PytorchClas
             Sample weights or None.
         early_stopping : EarlyStoppingHandler or 'none'
             A strategy for early stopping. Passing 'none' disables early stopping.
-        model_selection : ModelSelectionHandler or 'none'
+        model_selection : ModelSelectionHandler or None, default=None
             A model selection handler. Passing 'none' disables model selection.
         optimizer : torch.optim.optimizer.Optimizer or None, default=None
             A pytorch optimizer.
@@ -392,7 +393,9 @@ class TransformerBasedClassification(TransformerBasedEmbeddingMixin, PytorchClas
         with tempfile.TemporaryDirectory(dir=get_tmp_dir_base()) as tmp_dir:
             self._train(sub_train, sub_valid, weights, early_stopping, model_selection,
                         optimizer, scheduler, tmp_dir)
-            self._perform_model_selection(optimizer, model_selection)
+
+            if not isinstance(model_selection, NoopModelSelection):
+                self._perform_model_selection(optimizer, model_selection)
 
         return self
 
@@ -518,8 +521,9 @@ class TransformerBasedClassification(TransformerBasedEmbeddingMixin, PytorchClas
 
                     measured_values = {'val_loss': valid_loss, 'val_acc': valid_acc}
                     stop = stop or early_stopping.check_early_stop(num_epoch+1, measured_values)
-                    self._save_model(optimizer, model_selection, f'{num_epoch}-b{i+1}',
-                                     train_acc, train_loss, valid_acc, valid_loss, stop, tmp_dir)
+                    if not isinstance(model_selection, NoopModelSelection):
+                        self._save_model(optimizer, model_selection, f'{num_epoch}-b{i+1}',
+                                         train_acc, train_loss, valid_acc, valid_loss, stop, tmp_dir)
 
         if validate_every:
             valid_loss, valid_acc = np.mean(valid_losses), np.mean(valid_accs)
@@ -538,8 +542,10 @@ class TransformerBasedClassification(TransformerBasedEmbeddingMixin, PytorchClas
             'val_acc': valid_acc
         }
         stop = early_stopping.check_early_stop(num_epoch+1, measured_values)
-        self._save_model(optimizer, model_selection, f'{num_epoch}-b0',
-                         train_acc, train_loss, valid_acc, valid_loss, stop, tmp_dir)
+        if not isinstance(model_selection, NoopModelSelection):
+            self._save_model(optimizer, model_selection, f'{num_epoch}-b0',
+                             train_acc, train_loss, valid_acc, valid_loss, stop, tmp_dir)
+
         return train_loss, train_acc, valid_loss, valid_acc, stop
 
     def _create_collate_fn(self, use_sample_weights=False):